requirements.txt
requests
pandas
pytrends
python-dotenv
google-api-python-client
apify-client